moved server helper functions into separate file
